chore: prepare Release 38 for release testing
