+++ meteor/server/migration/X_X_X.ts
@@ -1,7 +1,5 @@
 import { addMigrationSteps } from './databaseMigration'
 import { CURRENT_SYSTEM_VERSION } from './currentSystemVersion'
-import { TriggeredActions } from '../../lib/collections/TriggeredActions'
-import { getHash, protectString, unprotectString } from '../../lib/lib'
 
 /*
  * **************************************************************************************
@@ -15,25 +13,4 @@ import { getHash, protectString, unprotectString } from '../../lib/lib'
 // Release X
 export const addSteps = addMigrationSteps(CURRENT_SYSTEM_VERSION, [
 	// Add some migrations!
-	{
-		id: `TriggeredActions.core.fixIds`,
-		canBeRunAutomatically: true,
-		validate: () => {
-			const existingActions = TriggeredActions.find({ showStyleBaseId: null }).fetch()
-			return existingActions.some((action) => !!unprotectString(action._id).match(/^core_/))
-		},
-		migrate: () => {
-			const existingActions = TriggeredActions.find({ showStyleBaseId: null }).fetch()
-			for (const action of existingActions) {
-				const actionId = unprotectString(action._id)
-				if (actionId.match(/^core_/) !== null) {
-					TriggeredActions.remove(action._id)
-					TriggeredActions.insert({
-						...action,
-						_id: protectString(getHash(actionId)),
-					})
-				}
-			}
-		},
-	},
 ])
