+++ meteor/server/migration/migrations.ts
@@ -108,6 +108,9 @@ addSteps1_36_0()
 import { addSteps as addSteps1_37_0 } from './1_37_0'
 addSteps1_37_0()
 
+import { addSteps as addSteps1_38_0 } from './1_38_0'
+addSteps1_38_0()
+
 // Migrations for the in-development release:
 import { addSteps as addStepsX_X_X } from './X_X_X'
 addStepsX_X_X()
